+++ gween.go
@@ -9,12 +9,12 @@ type (
 	// Tween encapsulates the easing function along with timing data. This allows
 	// a ease.TweenFunc to be used to be easily animated.
 	Tween struct {
-		duration float32
-		time     float32
-		begin    float32
-		end      float32
-		change   float32
-		Overflow float32
+		duration float64
+		time     float64
+		begin    float64
+		end      float64
+		change   float64
+		Overflow float64
 		easing   ease.TweenFunc
 		reverse  bool
 	}
@@ -24,7 +24,7 @@ type (
 // of the tween and the easing function to animate between the two values. The
 // easing function can be one of the provided easing functions from the ease package
 // or you can provide one of your own.
-func New(begin, end, duration float32, easing ease.TweenFunc) *Tween {
+func New(begin, end, duration float64, easing ease.TweenFunc) *Tween {
 	return &Tween{
 		begin:    begin,
 		end:      end,
@@ -38,7 +38,7 @@ func New(begin, end, duration float32, easing ease.TweenFunc) *Tween {
 
 // Set will set the current time along the duration of the tween. It will then return
 // the current value as well as a boolean to determine if the tween is finished.
-func (tween *Tween) Set(time float32) (current float32, isFinished bool) {
+func (tween *Tween) Set(time float64) (current float64, isFinished bool) {
 	if time <= 0 {
 		tween.Overflow = time
 		tween.time = 0
@@ -70,7 +70,7 @@ func (tween *Tween) Reset() {
 
 // Update will increment the timer of the Tween and ease the value. It will then
 // return the current value as well as a bool to mark if the tween is finished or not.
-func (tween *Tween) Update(dt float32) (current float32, isFinished bool) {
+func (tween *Tween) Update(dt float64) (current float64, isFinished bool) {
 	if tween.reverse {
 		return tween.Set(tween.time - dt)
 	}
